Fix the kdm init script to use any overrides set in /etc/default/kdm.d/.
Because default_blue.jpg is no longer the default wallpaper, checking
for it before using the overrides effectively made the overrides be ignored.
This new approach allows the overrides to be used without checking if the
default wallpaper is set in backgroundrc and only check it for overriding the
wallpaper itself, i.e. if the default wallpaper is set in backgroundrc, use
the overriden wallpaper (if any), otherwise use whatever is set.