Tie autostart handling to the explicit --desktop option in plasma-netbook.
This change was done in order to maintain command line option compatibility
with unpatched version of the binary. --desktop is supported by original
plasma-netbook where it has no effect while --autostart would trigger a failure
to start if a user used the same profile (and hence incompatible autostart
file) with original version. We don't want to leave a user without a plasma
shell.